Refactor slack-related module structure
